app/backend/requirements.txt
flask
opencv-python
ultralytics
python-dotenv
